sensor_server/sensor_integrate_v2.0/requirements.txt
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
future==0.18.2
iso8601==1.0.2
keyboard==0.13.5
numpy==1.23.0
psutil==5.9.1
pyserial==3.5
PyYAML==6.0
wincertstore==0.2
